% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addMaxNonPeakGene}
\alias{addMaxNonPeakGene}
\title{Calculate highest read outside peak (Genewise) Depends on the results of addMaxNonPeakRegion}
\usage{
addMaxNonPeakGene(pydeg_np_region, dup_ID, core = 1)
}
\arguments{
\item{pydeg_np_region}{data.frame with non-peak regions and max reads within them}

\item{dup_ID}{tx_names from transcripts to select the max read from}

\item{core}{Number of cores for this function}
}
\description{
Calculate highest read outside peak (Genewise) Depends on the results of addMaxNonPeakRegion
}
