% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addMaxNonPeakRegion}
\alias{addMaxNonPeakRegion}
\title{Calculate highest read outside peak (Adapted from the addMaxNonPeakSignal function)}
\usage{
addMaxNonPeakRegion(df, bigwig, i.comp, core = env$core)
}
\arguments{
\item{df}{data.frame (subset) of entries with more than one peak in the gene}

\item{bigwig}{Filename of the bigwig file}

\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{core}{Number of cores available}
}
\description{
Calculate highest read outside peak (Adapted from the addMaxNonPeakSignal function)
}
