% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addMaxNonPeakSignal}
\alias{addMaxNonPeakSignal}
\title{Add maxNonPeakRegion to pyDegradome result data.frame}
\usage{
addMaxNonPeakSignal(i.df, i.comp, core = 1)
}
\arguments{
\item{core}{Number of cores available}

\item{x}{data.frame of pyDegradome result}

\item{ann}{data.frame of GFF annotation}
}
\description{
Add maxNonPeakRegion to pyDegradome result data.frame
}
