% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addNonPeak}
\alias{addNonPeak}
\title{Calculate highest peak read (Adapted from the addMaxNonPeakSignal function)}
\usage{
addNonPeak(df_dup_ID_sub, df_dup_sub, bigwig, i.comp, core = env$cores)
}
\arguments{
\item{df_dup_ID_sub}{data.frame (subset) of entries with more than one peak in the gene}

\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{core}{Number of cores available}
}
\description{
Calculate highest peak read (Adapted from the addMaxNonPeakSignal function)
}
