% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{checkNonPeakRefOne}
\alias{checkNonPeakRefOne}
\title{Calculate highest read outside peak (peak = 1)}
\usage{
checkNonPeakRefOne(df, i.comp, cols.indx, idir = pydeg_reads_dir)
}
\arguments{
\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{cols.indx}{A vector of column names used to construct the index.}

\item{i.df}{data.frame of pyDegradome result}
}
\description{
Calculate highest read outside peak (peak = 1)
}
