% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_profiles.R
\name{getAbsCountsDistList}
\alias{getAbsCountsDistList}
\title{Variation of getRelativeMeanDist. It also outputs the raw counts as a data frame. Information about the transcript is collected. This was written to get the transcript id for those peaks of interest}
\usage{
getAbsCountsDistList(x, l = -50, thr = 1)
}
\arguments{
\item{x}{Numeric data.frame or matrix.}

\item{thr}{Numeric. Threshold for total count. Records whose total counts are less than `thr` will be excluded from the calculation.}
}
\value{
List: data frames of raw counts that have at least one read over the range (-50:50)
}
\description{
Variation of getRelativeMeanDist. It also outputs the raw counts as a data frame. Information about the transcript is collected. This was written to get the transcript id for those peaks of interest
}
