% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{getFirstComparison}
\alias{getFirstComparison}
\title{Takes the name of a pydegradome comparison ('t_<comparison_1>_c_<comparison_2>') and returns the first element (<comparison_1>).}
\usage{
getFirstComparison(i.comp)
}
\arguments{
\item{icomp}{filename based on the comparison between test and control samples}
}
\description{
Takes the name of a pydegradome comparison ('t_<comparison_1>_c_<comparison_2>') and returns the first element (<comparison_1>).
}
