% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{getMaxNonPeak_multiple}
\alias{getMaxNonPeak_multiple}
\title{Takes the output of getMaxNonPeakRegion and return the highest read
along the gene (loops through isoforms)}
\usage{
getMaxNonPeak_multiple(pydeg_np_region, dup_ID, core = 1)
}
\arguments{
\item{pydeg_np_region}{Ouput from getMaxNonPeakRegion}

\item{dup_ID}{tx_names from transcripts to select the max read from}

\item{core}{Number of cores for this function}
}
\description{
Takes the output of getMaxNonPeakRegion and return the highest read
along the gene (loops through isoforms)
}
