% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{getMaxNonPeak_single}
\alias{getMaxNonPeak_single}
\title{Takes a dataframe with coordinates for a single peak,
gene start and end positions. With those coordinates it process a bigwig file
and gets highest read value outside the peak region.
Results are appended to the input dataframe}
\usage{
getMaxNonPeak_single(df, i.comp, cols.indx, idir = pydeg_reads_dir)
}
\arguments{
\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{cols.indx}{A vector of column names used to construct the index.}

\item{i.df}{data.frame of pyDegradome result}
}
\description{
Takes a dataframe with coordinates for a single peak,
gene start and end positions. With those coordinates it process a bigwig file
and gets highest read value outside the peak region.
Results are appended to the input dataframe
}
