% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{getMaxPeak}
\alias{getMaxPeak}
\title{Takes a dataframe with coordinates for a peak region and calculates
from a bigwig file the highest read value for the peak.
Results are appended to the input dataframe}
\usage{
getMaxPeak(df, i.comp, i.cols = cols.indx, idir = pydeg_reads_dir)
}
\arguments{
\item{df}{data.frame of pyDegradome result}

\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{i.cols}{vector of columns to subset the table to append to the reference file}

\item{idir}{path to the directory with the reference file}
}
\description{
Takes a dataframe with coordinates for a peak region and calculates
from a bigwig file the highest read value for the peak.
Results are appended to the input dataframe
}
