% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_visualization.R
\name{GetPlotCoors}
\alias{GetPlotCoors}
\title{Returns: Start and end coordinates for the plot}
\usage{
GetPlotCoors(i_start, i_end, i_factor = NULL, p_width = NULL)
}
\arguments{
\item{i_start}{ranscript start coordinates}

\item{i_end}{Transcript end coordinates}

\item{i_factor}{proportion of the transcript range that will be added as offset}

\item{p_width}{width of the plot in basepairs}
}
\description{
Returns: Start and end coordinates for the plot
}
