% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_visualization.R
\name{GetPlotMarks}
\alias{GetPlotMarks}
\title{Returns: Coordinates on the GenomeAxisTrack for the marks based on an intial number of segments (ticksn)}
\usage{
GetPlotMarks(
  i_start,
  i_end,
  p_start,
  p_end,
  iticks,
  txPlot = FALSE,
  min_ticks = 4,
  max_ticks = 6
)
}
\arguments{
\item{p_start}{ranscript start coordinates}

\item{p_end}{Transcript end coordinates}

\item{ticksn}{Number of segments to divide the transcript length}

\item{i_factor}{proportion of the transcript range that will be added as offset}

\item{p_width}{width of the plot in basepairs}
}
\description{
Returns: Coordinates on the GenomeAxisTrack for the marks based on an intial number of segments (ticksn)
}
