% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_profiles.R
\name{getProfileFromGrange}
\alias{getProfileFromGrange}
\title{Extract profiles of specific regions from a list of matrixes of profiles.}
\usage{
getProfileFromGrange(psite_prof, gr)
}
\arguments{
\item{psite_prof}{`psite_prof` object (list[by transcript] of matrix[read count of position x sample]) from Ribo-seq analysis.}

\item{gr}{`Granges` to define the area of interest.}
}
\value{
List[by region] of matrix[read count of position x sample].
}
\description{
Extract profiles of specific regions from a list of matrixes of profiles.
}
