% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_profiles.R
\name{getRelativeMeanDist}
\alias{getRelativeMeanDist}
\title{Calculate mean of relative distribution from a matrix or data.frame of counts(position x count).}
\usage{
getRelativeMeanDist(x, thr = 1)
}
\arguments{
\item{x}{Numeric data.frame or matrix.}

\item{thr}{Numeric. Threshold for total count. Records whose total counts are less than `thr` will be excluded from the calculation.}
}
\value{
Numeric vector of normalized mean count. The last value is the number of exon counted.
}
\description{
Calculate mean of relative distribution from a matrix or data.frame of counts(position x count).
}
