% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_profiles.R
\name{getRelativeMeanDist2}
\alias{getRelativeMeanDist2}
\title{Variation of getRelativeMeanDist. It calculates the mean distribution while checking that the input matrix is not empty}
\usage{
getRelativeMeanDist2(x, thr = 1)
}
\arguments{
\item{x}{Numeric data.frame or matrix.}

\item{thr}{Numeric. Threshold for total count. Records whose total counts are less than `thr` will be excluded from the calculation.}
}
\value{
List of numeric vectors of normalized mean counts. The last value is the number of exon counted.
}
\description{
Variation of getRelativeMeanDist. It calculates the mean distribution while checking that the input matrix is not empty
}
