% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_visualization.R
\name{invertGRscores}
\alias{invertGRscores}
\title{Invert the score values of a list of a GenomicRange object}
\usage{
invertGRscores(bwl)
}
\arguments{
\item{bwl}{A list of GenomicRange-objects with a score metadata}
}
\description{
Invert the score values of a list of a GenomicRange object
}
