% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_visualization.R
\name{invertVector}
\alias{invertVector}
\title{Invert values and order of elements of a numeric vector while keeping names}
\usage{
invertVector(x)
}
\arguments{
\item{x}{A numeric vector with names}
}
\description{
Invert values and order of elements of a numeric vector while keeping names
}
