% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MatchDFwithRef}
\alias{MatchDFwithRef}
\title{Merges a column (ref_col) from a reference dataframe (i.reads) into an input dataframe (df). An 'indx' is used for merging. The input is divided into matching and non-matching dataframes and returned as a list.}
\usage{
MatchDFwithRef(df, i.reads, ref_col)
}
\arguments{
\item{df}{data.frame to match against a reference}

\item{i.reads}{reference dataframe}

\item{ref_col}{Column with values to merge}
}
\description{
Merges a column (ref_col) from a reference dataframe (i.reads) into an input dataframe (df). An 'indx' is used for merging. The input is divided into matching and non-matching dataframes and returned as a list.
}
