% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_visualization.R
\name{maxCoverageBigWigList}
\alias{maxCoverageBigWigList}
\title{Get max and min coverage from a BigWig file for specific region.}
\usage{
maxCoverageBigWigList(bigwigs, chr, x_start, x_end)
}
\arguments{
\item{bigwigs}{List of bigwig records. BigWig files should be loaded by `rtracklayer::import`.}

\item{chr}{Char of chromosome name}

\item{x_start}{Numeric. Coordinate of left end of the plot range.}

\item{x_end}{Numeric. Coordinate of right end of the plot range.}
}
\description{
Get max and min coverage from a BigWig file for specific region.
}
