% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MaxPeakBigWig}
\alias{MaxPeakBigWig}
\title{Takes a data frame and a comparison and returns the highest read using GetMaxread}
\usage{
MaxPeakBigWig(df, i.comp, core = 1)
}
\arguments{
\item{df}{data.frame with PyDegradome result}

\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{core}{Number of threads for the parallel process}
}
\description{
Takes a data frame and a comparison and returns the highest read using GetMaxread
}
