% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{maxPtest}
\alias{maxPtest}
\title{Get the mean value of peak between replicates in the test sample}
\usage{
maxPtest(f_df, f_bigwigs, f_pairs, f_input_dir, f_core = 1)
}
\arguments{
\item{f_df}{data frame with pydegradome processed output}

\item{f_bigwigs}{bigwigs List of bigwig records. BigWig files should be loaded by `rtracklayer::import}

\item{f_pairs}{pairs of genotype B concentration to be compared}

\item{f_input_dir}{directory of the pooled file}

\item{f_core}{Number of threads for the nested funtion (GetMaxread)}
}
\description{
Get the mean value of peak between replicates in the test sample
}
