% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{maxReadPairs}
\alias{maxReadPairs}
\title{Takes a dataframe with start and stop coordinates and calculates (from a bigwig file)
the highest read value. The function loops over a list of bigwig files
corresponding to replicates of a particular condition (test or control). The boolean 'gene' is used to define the columns used to get coordinates and the name of the reference file If TRUE, coordinates will be taken from the gene_region_ column and the reference will be have the root name 'Max_reads_tx-'; if FALSE coordinates}
\usage{
maxReadPairs(f_df, f_bigwigs, f_pairs, f_input_dir, gene = FALSE, f_core = 1)
}
\arguments{
\item{f_df}{data frame with pydegradome processed output}

\item{f_bigwigs}{List of bigwig records. BigWig files should be loaded by `rtracklayer::import}

\item{f_pairs}{pairs of genotype B concentration to be compared}

\item{f_input_dir}{directory of the pooled file}

\item{gene}{boolean used to define which columns and references will be used. TRUE to calculate control max read and FALSE to calculate peak max read (test)}

\item{f_core}{Number of threads for the nested funtion (getMaxread)}
}
\description{
Takes a dataframe with start and stop coordinates and calculates (from a bigwig file)
the highest read value. The function loops over a list of bigwig files
corresponding to replicates of a particular condition (test or control). The boolean 'gene' is used to define the columns used to get coordinates and the name of the reference file If TRUE, coordinates will be taken from the gene_region_ column and the reference will be have the root name 'Max_reads_tx-'; if FALSE coordinates
}
