% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MergeDFs}
\alias{MergeDFs}
\title{Merges a dataframe with intermediary results into the main dataframe}
\usage{
MergeDFs(dfA, dfB, ref_col)
}
\arguments{
\item{dfA}{Dataframe to merge into}

\item{dfB}{Dataframe to merge from}

\item{ref_col}{Column name (string) to merge}
}
\description{
Merges a dataframe with intermediary results into the main dataframe
}
