% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_visualization.R
\name{mergeVector}
\alias{mergeVector}
\title{Invert values and order of elements of a numeric vector while keeping names}
\usage{
mergeVector(
  ivec,
  sep1 = " vs ",
  sep2 = " AND \\n",
  parenthesis = c("(", ")"),
  single = FALSE
)
}
\arguments{
\item{ivec}{A string vector with names to construct the output}

\item{sep1}{A string to separate the first half of vector elements}

\item{sep2}{A string to separate the two halfs before exporting}

\item{parenthesis}{A pair of symbols to group the half of vector elements}
}
\description{
Invert values and order of elements of a numeric vector while keeping names
}
