% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_profiles.R
\name{omitLastExon}
\alias{omitLastExon}
\title{Omit last record from a Granges object considering the strandness.}
\usage{
omitLastExon(x)
}
\arguments{
\item{x}{`Granges` object of exons in a transcript}
}
\value{
`Granges` object with lenght of `length(x) - 1`, after omiting the last (3'end in transcript) record in the transcript
}
\description{
Omit last record from a Granges object considering the strandness.
}
