% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{plotProfilesAllPydegResults}
\alias{plotProfilesAllPydegResults}
\title{Save pdf of degradome profiles around detected peaks from a processed pyDegradome result table.}
\usage{
plotProfilesAllPydegResults(
  pydeg,
  sample_list,
  bigwigs,
  bigwigs2 = NULL,
  out_dir,
  min_plot_width = 1000,
  col_cond = "blue",
  col_hl = "#00ff0099",
  plotrange = "peak",
  tr_range = NULL,
  core = 1
)
}
\arguments{
\item{pydeg}{data.frame of (processed) pyDegradome output.}

\item{sample_list}{Character vector of sampel names to be plotted. Should be a part of `names(bigwigs)`.}

\item{bigwigs}{List of bigwig records. BigWig files should be loaded by `rtracklayer::import`.}

\item{out_dir}{Path to output directory.}

\item{min_plot_width}{Numeric. Minimum plotting range.}

\item{col_cond}{Named vector of color code. This must have names identical to `names(bigwigs)`.}

\item{col_hl}{Charactor (vector) of color code to be used in highlighting.}

\item{plotrange}{Character. Define range of the plot.'transcript' for whole transcript or 'peak' for around peaks. `tr_range` must be given for 'transcript'}

\item{tr_range}{Granges object for transcript ranges. Result of `GenomicFeatures::transcripts(txdb)`}

\item{core}{Number of thread for multi-threading.}
}
\description{
Save pdf of degradome profiles around detected peaks from a processed pyDegradome result table.
}
