% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{SampleName}
\alias{SampleName}
\title{Gets the first element of the comparison used for the pydegradome analysis}
\usage{
SampleName(i.comp)
}
\arguments{
\item{icomp}{filename based on the comparison between test and control samples}
}
\description{
Gets the first element of the comparison used for the pydegradome analysis
}
