% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{splitDFbyCol}
\alias{splitDFbyCol}
\title{Splits a dataframe based on whether the values of a column are duplicated or unique.
Returns a list of dataframes}
\usage{
splitDFbyCol(df, ref_col, keep_ref = FALSE)
}
\arguments{
\item{df}{Dataframe with peak and gene coordinates}

\item{ref_col}{Column used to judge duplicates}

\item{keep_ref}{Boolean for keeping or removing 'ref_col'}
}
\description{
Splits a dataframe based on whether the values of a column are duplicated or unique.
Returns a list of dataframes
}
