% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{SplitDFbyNpeaks}
\alias{SplitDFbyNpeaks}
\title{Find genes with two peaks based on repetition of the ID value}
\usage{
SplitDFbyNpeaks(df, ref_col)
}
\arguments{
\item{df}{Dataframe with peak and gene coordinates}

\item{ref_col}{Column used to judge duplicates}
}
\description{
Find genes with two peaks based on repetition of the ID value
}
