% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addMaxNonPeakGene}
\alias{addMaxNonPeakGene}
\title{Takes the output of addMaxNonPeakRegion and return the highest read
along the gene (loops through isoforms)}
\usage{
addMaxNonPeakGene(pydeg_np_region, dup_ID, core = 1)
}
\arguments{
\item{pydeg_np_region}{Ouput from addMaxNonPeakRegion}

\item{dup_ID}{tx_names from transcripts to select the max read from}

\item{core}{Number of cores for this function}
}
\description{
Takes the output of addMaxNonPeakRegion and return the highest read
along the gene (loops through isoforms)
}
