% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addMaxNonPeakRegion}
\alias{addMaxNonPeakRegion}
\title{Takes a data frame with non peak coordinates and
returns the highest read per region}
\usage{
addMaxNonPeakRegion(df, bigwig, i.comp, core = env$cores)
}
\arguments{
\item{df}{A data.frame with non peak coordinates of entries with more than one peak in the gene}

\item{bigwig}{Filename of the bigwig file}

\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{core}{Number of threads for the parallel process}
}
\description{
Takes a data frame with non peak coordinates and
returns the highest read per region
}
