% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{addMaxNonPeakSignal}
\alias{addMaxNonPeakSignal}
\title{Returns the highest read outside the peak region.
Considers regions upstream and downstream of the peak}
\usage{
addMaxNonPeakSignal(i.df, i.comp, core = 1)
}
\arguments{
\item{i.df}{data frame with PyDegradome results}

\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{core}{Number of threads for the parallel process}
}
\description{
Returns the highest read outside the peak region.
Considers regions upstream and downstream of the peak
}
