% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{checkPeakRef}
\alias{checkPeakRef}
\title{Calculate reads for a given region. Checks a reference file to speed up}
\usage{
checkPeakRef(df, i.comp, i.cols = cols.indx, idir = pydeg_reads_dir)
}
\arguments{
\item{df}{data.frame of pyDegradome result}

\item{i.comp}{comparison}
}
\description{
Calculate reads for a given region. Checks a reference file to speed up
}
