% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{getMaxRead}
\alias{getMaxRead}
\title{Reads a bigwig file and returns a vector with the highest read value over a set of genomic ranges (one or two GRanges objects).}
\usage{
getMaxRead(grA, f_bigwig, grB = NULL, core = 1)
}
\arguments{
\item{f_bigwig}{reference to a bigwig file}

\item{core}{Number of threads for the parallel process}

\item{gr_A}{main GRanges object}

\item{gr_B}{secondary GRanges object}
}
\description{
Reads a bigwig file and returns a vector with the highest read value over a set of genomic ranges (one or two GRanges objects).
}
