% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{GetMaxRead}
\alias{GetMaxRead}
\title{Calculates the highest read value from a series of GRanges using a foreach loop}
\usage{
GetMaxRead(grA, f_bigwig, grB = NULL, core = 1)
}
\arguments{
\item{f_bigwig}{reference to a bigwig file}

\item{core}{Number of threads for the parallel process}

\item{gr_A}{main GRanges object}

\item{gr_B}{secondary GRanges object}
}
\description{
Calculates the highest read value from a series of GRanges using a foreach loop
}
