% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{GetMaxRead}
\alias{GetMaxRead}
\title{Calculates the max read in a series of GRanges using a foreach loop}
\usage{
GetMaxRead(grA, f_bigwig, grB = NULL, core = 1)
}
\arguments{
\item{gr}{GRanges object}

\item{bigwig}{reference to a bigwig file}
}
\description{
Calculates the max read in a series of GRanges using a foreach loop
}
