% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{getNonPeakCoors}
\alias{getNonPeakCoors}
\title{Returns a list of coordinates for non peak regions.}
\usage{
getNonPeakCoors(df)
}
\arguments{
\item{df}{A dataframe with peak and gene coordinates}
}
\description{
Returns a list of coordinates for non peak regions.
}
