% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MatchDFwithRef}
\alias{MatchDFwithRef}
\title{Merges a dataframe (df) with a particular column (ref_col) reference based on the index 'indx' column
Returns a list of dataframes one matched with the reference an another with nonmatches}
\usage{
MatchDFwithRef(df, i.reads, ref_col)
}
\arguments{
\item{df}{data.frame to match against a reference}

\item{i.reads}{reference dataframe}

\item{ref_col}{Column with values to merge}
}
\description{
Merges a dataframe (df) with a particular column (ref_col) reference based on the index 'indx' column
Returns a list of dataframes one matched with the reference an another with nonmatches
}
