% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MatchDFwithRef}
\alias{MatchDFwithRef}
\title{Merges a given column of a dataframe with a reference based on an index 'indx' column
Returns a list of dataframes one matched with the reference an another with nonmatches}
\usage{
MatchDFwithRef(df, i.reads, ref_col)
}
\arguments{
\item{df}{data.frame to match against a reference}

\item{ref_col}{Column with values to merge}

\item{ref}{reference dataframe}
}
\description{
Merges a given column of a dataframe with a reference based on an index 'indx' column
Returns a list of dataframes one matched with the reference an another with nonmatches
}
