% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MaxPeakBigWig}
\alias{MaxPeakBigWig}
\title{Calculate highest peak read (Adapted from the addMaxNonPeakSignal function)}
\usage{
MaxPeakBigWig(df, i.comp, core = 1)
}
\arguments{
\item{i.comp}{Name of comparison pair (divided into test and ctrl)}

\item{core}{Number of cores available}

\item{i.df}{data.frame of pyDegradome result}
}
\description{
Calculate highest peak read (Adapted from the addMaxNonPeakSignal function)
}
