% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{maxTxctrl}
\alias{maxTxctrl}
\title{Get the mean value of the highest signal found in each of the replicates in control samples}
\usage{
maxTxctrl(tx_id, f_bigwigs, f_pairs, f_gr_tr, f_input_dir, f_core = 1)
}
\arguments{
\item{tx_id}{Vector with transcripts from which the reads should be obtained}

\item{f_bigwigs}{bigwigs List of bigwig records. BigWig files should be loaded by `rtracklayer::import}

\item{f_pairs}{pairs of genotype B concentration to be compared}

\item{f_gr_tr}{subset of the GRanges object tx_range using pydeg_tx}

\item{f_input_dir}{directory of the pooled file}

\item{f_core}{Number of threads for the nested function (GetMaxread)}
}
\description{
Get the mean value of the highest signal found in each of the replicates in control samples
}
