% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MergeDFs}
\alias{MergeDFs}
\title{Merges two data frames by an indx column composed by tx_name, chr and strand}
\usage{
MergeDFs(dfA, dfB, ref_col)
}
\arguments{
\item{dfA}{Dataframe to merge into}

\item{dfB}{Dataframe to merge from}

\item{ref_col}{Column name (string) to merge}
}
\description{
Merges two data frames by an indx column composed by tx_name, chr and strand
}
