% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{MergeDFs}
\alias{MergeDFs}
\title{Merges a column (ref_col) from dataframe B (dfB) into dataframe A (dfA) keeping all entries from the latter. Returns a dataframe}
\usage{
MergeDFs(dfA, dfB, ref_col)
}
\arguments{
\item{dfA}{Dataframe to merge into}

\item{dfB}{Dataframe to merge from}

\item{ref_col}{Column name (string) to merge}
}
\description{
Merges a column (ref_col) from dataframe B (dfB) into dataframe A (dfA) keeping all entries from the latter. Returns a dataframe
}
