% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pydegradome.R
\name{SampleName}
\alias{SampleName}
\title{Takes the filename of the pydegradome output and returns the first element (first comparison)}
\usage{
SampleName(i.comp)
}
\arguments{
\item{icomp}{filename based on the comparison between test and control samples}
}
\description{
Takes the filename of the pydegradome output and returns the first element (first comparison)
}
